Chore: organize code
코드 정리 및 일부 기능 업데이트
- 스테이지 정보 저장시 파일 브라우저 팝업을 사용하도록 수정
