--- Assets/ScriptableObjects/TestStage.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d957c1d2e878c014e9fcd926e41bf6e3
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
