+++ Assets/Scripts/StageBuilder/PathStruct.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ef60900e59a6344dafd569e22b5ef34
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
