+++ Assets/Scripts/StageBuilder/TileNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: baba44e0c53fcbf44b7f61c523b0543a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
